renaming main file should have been done with 12a1a28, I think.
