+++ lib/orientdb-schema-migrator.rb
