+++ spec/spec_helper.rb
@@ -19,7 +19,7 @@
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), "..", "lib"))
 
-require 'orientdb_schema_migrator'
+require 'orientdb-schema-migrator'
 require 'climate_control'
 
 RSpec.configure do |config|
